refactor(extension_host): extract extension module, fix Registry typo
Move extension module from daemon/core to ozmux_extension_host crate.
Fix ExtensionRegistory → ExtensionRegistry typo. Change serve() to
accept SessionState instead of AppState.
