+++ daemon/extension_host/src/manifest.rs
@@ -2,7 +2,7 @@ use std::{collections::HashMap, path::PathBuf};
 
 use serde::{Deserialize, Serialize};
 
-use crate::define_string_new_type;
+use ozmux_macros::define_string_new_type;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct ExtensionManifest {
